refactor: delete the superseded manifest.Extractor and its unfixed traversal
pkg/manifest/extract.go held the original selective-extraction implementation
from #93. It has no non-test caller, and it carried an unfixed copy of the path
traversal that #282 fixed elsewhere:

    // getOutputPath, extract.go:413
    return filepath.Join(req.OutputDir, filePath)   // filePath is a tar header.Name

The result flowed straight into os.MkdirAll + os.Create, so a chunk containing
an entry named ../../../etc/cron.d/x would write outside OutputDir. Both live
extraction paths guard against this — SelectiveExtractor.restorePath (fixed in
#282, fuzzed by FuzzRestorePath) and pkg/extraction's explicit containment
check — but this third copy never got the fix.

Not exploitable in any released version: every user-facing restore path
(restore, download, browse, shell, restore_jobs, the TUI) constructs a
SelectiveExtractor. The risk was that this is exported API in pkg/, so an
external importer, or a future internal caller reaching for the more
familiar-sounding NewExtractor, would silently get the unsafe extractor.

Deleted rather than patched. git log shows extract.go untouched since #207 (a
CI repair) while SelectiveExtractor absorbed every correctness and trust fix
(#229, #281, #282, #283, #287). Fixing getOutputPath would mean maintaining a
second sanitizer and a second extraction loop that nothing calls — exactly the
duplication that let #282 survive in one copy after being fixed in the other.

Removed: Extractor, NewExtractor, Extract, EstimateExtractCost,
ListExtractableFiles, ExtractRequest, ExtractResult, and the private helpers.

Kept S3Downloader, which happened to be declared in this file and is used
repo-wide (cmd/verify.go, pkg/tui/browser.go, the dvc integration and benchmark
suites); it moves to pkg/manifest/s3.go. The three test fixtures that outlived
their own file — mockS3Client, createTestManifest, createMockChunks, still used
by batch_restore_test.go and direct_restore_test.go — move verbatim to
helpers_test.go.

The deleted code was better covered than the package average (~85% vs 77.1%),
so pkg/manifest drops to 76.5%. That is still above its .coverage-baseline
floor of 75, so the baseline is unchanged; propping up a ratio was not a reason
to keep dead code.

Closes #308
